+++ man-roxygen/DbmsDetails.R
@@ -15,11 +15,6 @@
 #' @param password           The password for that user.
 #' @param server             The name of the server.
 #' @param port               (optional) The port on the server to connect to.
-#' @param schema             (DEPRECATED. optional) The name of the schema to connect to. 
-#'                           DatabaseConnector attempts to set the schema using USE statement, but
-#'                           not all DBMS support USE statements. In DBMS that don't support USE
-#'                           statements this strategy will fail. Instead use a fully specified table 
-#'                           names in your SQL statement, for example 'SELECT * FROM my_schema.my_table;'
 #' @param extraSettings      (optional) Additional configuration settings specific to the database
 #'                           provider to configure things as security for SSL. These must follow the
 #'                           format for the JDBC connection for the RDBMS specified in dbms.
@@ -43,8 +38,6 @@
 #'   \item \code{server}. This field contains the SID, or host and servicename, SID, or TNSName:
 #'         '<sid>', '<host>/<sid>', '<host>/<service name>', or '<tnsname>'
 #'   \item \code{port}. Specifies the port on the server (default = 1521)
-#'   \item \code{schema}. DEPRECATED: This field contains the schema (i.e. 'user' in Oracle terms) containing the
-#'         tables
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "(PROTOCOL=tcps)")
 #'   \item \code{oracleDriver} The driver to be used. Choose between "thin" or "oci".
@@ -57,8 +50,6 @@
 #'   \item \code{password}. The password used to log on to the server
 #'   \item \code{server}. This field contains the host name of the server
 #'   \item \code{port}. Not used for SQL Server
-#'   \item \code{schema}. DEPRECATED: The database containing the tables. If both database and schema are specified
-#'         (e.g. 'my_database.dbo', then only the database part is used, the schema is ignored.
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "encrypt=true; trustServerCertificate=false;")
 #' }
@@ -70,7 +61,6 @@
 #'   \item \code{password}. The password used to log on to the server
 #'   \item \code{server}. This field contains the host name of the server
 #'   \item \code{port}. Not used for SQL Server
-#'   \item \code{schema}. DEPRECATED: The database containing the tables
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "encrypt=true; trustServerCertificate=false;")
 #' }
@@ -81,7 +71,6 @@
 #'   \item \code{server}. This field contains the host name of the server and the database holding the
 #'         relevant schemas: <host>/<database>
 #'   \item \code{port}. Specifies the port on the server (default = 5432)
-#'   \item \code{schema}. DEPRECATED: The schema containing the tables.
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "ssl=true")
 #' }
@@ -92,7 +81,6 @@
 #'   \item \code{server}. This field contains the host name of the server and the database holding the
 #'         relevant schemas: <host>/<database>
 #'   \item \code{port}. Specifies the port on the server (default = 5439)
-#'   \item \code{schema}. DEPRECATED: The schema containing the tables.
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "ssl=true&sslfactory=com.amazon.redshift.ssl.NonValidatingFactory")
 #' }
@@ -103,7 +91,6 @@
 #'   \item \code{server}. This field contains the host name of the server and the database holding the
 #'         relevant schemas: <host>/<database>
 #'   \item \code{port}. Specifies the port on the server (default = 5480)
-#'   \item \code{schema}. DEPRECATED: The schema containing the tables.
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "ssl=true")
 #'   \item \code{pathToDriver} The path to the folder containing the Netezza JDBC driver JAR file
@@ -115,7 +102,6 @@
 #'   \item \code{password}. The password for that user
 #'   \item \code{server}. The host name of the server
 #'   \item \code{port}. Specifies the port on the server (default = 21050)
-#'   \item \code{schema}. DEPRECATED: The database containing the tables
 #'   \item \code{extraSettings} The configuration settings for the connection (i.e. SSL Settings such
 #'         as "SSLKeyStorePwd=*****")
 #'   \item \code{pathToDriver} The path to the folder containing the Impala JDBC driver JAR files.