+++ NEWS.md
@@ -3,7 +3,7 @@ DatabaseConnector 4.0.0
 
 Changes:
 
-1. Support for 64-bit integers using BIGINT and bit64's integer64.
+1. Support for 64-bit integers using BIGINT and bit64's `integer64` type.
 
 2. INT fields are now translated to R integers and back.
 
@@ -12,6 +12,8 @@ Changes:
 4. Batch mode in executeSql divides SQL into batches (1000 statements per batch) to avoid running our of Java heap memory.
 
 5. ConnectionDetails delays evaluation of sensitive arguments until needed for improved security.
+
+6. Removing deprecated `schema` argument from `createConnectionDetails` and `connect`.
 
 
 Bugfixes: